+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.11.4
+
+### Patch Changes
+
+- [#193](https://github.com/wpsocio/wp-projects/pull/193) [`35f1abc`](https://github.com/wpsocio/wp-projects/commit/35f1abc0ad516ad2fa87bd4cbb3406fb77be1bc5) Thanks [@irshadahmad21](https://github.com/irshadahmad21)! - Performance improvements
+
+- [#195](https://github.com/wpsocio/wp-projects/pull/195) [`2c78dda`](https://github.com/wpsocio/wp-projects/commit/2c78ddac324fdb52ee47bb51fd293d5ddef18531) Thanks [@irshadahmad21](https://github.com/irshadahmad21)! - Fixed Telegram login redirect for Mini Apps
+
 ## 1.11.3
 
 ### Patch Changes